Automatically enable i2c_vc for imx290 if needed
This should automatically detect the imx290, check if i2c_vc is
enabled in config.txt, and if not enable it and immediately reboot.

This should prevent anyone from having to care about i2c_vc while also
making it possible to avoid turning it on by default (it breaks the
pi touchscreen and is a serious stability risk on the ground side)

Note: this is behind an intentionally undocumented setting for now,
to test it just put AUTO_I2C_VC=Y in openhd-settings-1.txt and boot
the air side with an imx290 connected. After we've tested this a bit
it can be enabled all the time without a setting.