Don't use red color on bitrate if set manually
If auto bitrate is in use, turning the bitrate red in the OSD is
helpful to indicate that real bitrate usage is approaching or has
exceeded the measured available bitrate.

However when manual bitrate is in use, there is nothing to indicate
what would be "too much" because there's no measurement to compare it
with in the first place.

At the moment the OSD is saying "the bitrate you configured is too
much" and that's not the case.

In the future we might want to still set this red based on some other
criteria, or always measure available bandwidth periodically.