Always use correct structure for RC packet layout
The old code was hardcoding the struct layout on the air side as if
the incoming RC packets were 13 bytes long and contained data for 16
switches, ignoring the JSSWITCHES setting in joyconfig.txt.

The ground side however was honoring the setting and only sending the
proper number of bits to the air side in RC packets.

The air side was therefore reading past the end of the packet in
memory and sending garbage values to the flight controller.

Further, the settings file claimed you could send either 8 or 16
switches, but this was totally impossible as we are always sending
8 axis channels, leaving only another 10 possible for Mavlink and none
of the other RC protocols were being given switch data at all.