Fix QGC heartbeat making OSD change flight mode
This is the same problem that affected QOpenHD, when you have QGC
running somewhere it sends its own heartbeats, which the OSD is
interpreting and using to change the flight mode and arm state shown
on the display.

This is only a quick fix, a real fix would be to validate the
heartbeat message.