Fix draw_TOTAL_DISTANCE, value is already KM/h
The telemetry code for each protocol is already standardizing the
value of the gps speed to KM/h so the 3.6 multiplication is already
being done elsewhere.

This changes the function signature to properly document the unit of
the variable.

This partially reverts 340b8aad858d807a31873a994b2ed4243703a27d

Closes: #263