Set up dedicated mavlink channel for microservices
This enables a separate mavlink communication channel between the
air and ground side, using OpenHDRouter at each end to transport
mavlink messages to/from the GCS (QOpenHD) and microservices running
on the air and ground pi.

The microservices are still under development but the GPIO service
and the power service both should be working. The camera service is
not yet implemented.