Allow different layer for hello_video rendering
On Pi4 with the v3d/vc4 driver, the video is on layer -1 so it is
rendering top of QOpenHD on layer -127.

However on existing OpenHD images, video is rendering on layer 0 and
the console is on layer -127, so we don't want to push it lower in
that case.

So we add a parameter to hello_video, if called as "hello_video 1",
the video will be pushed down to layer -128, otherwise it won't be.