Replace camera detection with python regex
We need to know whether there are one *or more* detected cameras,
because the stereo pi has 2 of them and other carriers or future pi
models might as well.

On the stereo pi, the old check fails because it says "detected=2".

This is still fragile as it depends on the output of vcgencmd, but
it doesn't seem to have ever changed.