Support boards that support multiple cameras
These scripts were checking to see if a camera is connected at all,
however they were assuming that the output of vcgencmd will contain
"supported=1" or "detected=1" when a camera is connected, which is
not always true.

On a stereo pi, supported=2 all the time, even with no camera
connected, and detected can be 0-2. On other boards, supported=1
most (all?) of the time.

To handle this, we use a regex which allows any integer to be in the
supported= part of the line, and then just capture the number in
detected=, whatever that may be.

We hope the output of vcgencmd won't change, but it would be easy to
spot if it happens.