Add support for automating frequency selection
We detect if an 8812 card is connected, in which case 5.8g band is
available for SmartSync and normal use so we prioritize it, otherwise
we use 2.4g.

If in any case we don't know what band is available or the settings
file doesn't explicitly say "auto" or a specific frequency, we default
to the same frequency we use by default anyway, 2412Mhz.