Automatically replace omxh264enc on buster
People have old settings files that work on stretch but won't on
buster, at least for now. The issue is that the raspbian buster
package for gst-omx does not work properly, and the only alternative
is v4l2h264enc.

We haven't yet determined the best replacement options for v4l2h264enc
but simply using the element by itself is enough to make things work
temporarily so that people don't think their camera setup is broken.

People will still need to replace the USBCamera line in their
settings file and tune it for their particular setup.