Ensure QOpenHD on ground receives FC telemetry
Ordinarily when a device joins the hotspot, a number of events take
place that cause telemetry streams to start forwarding to the device.

However, QOpenHD runs on the ground station itself, so there is no
"join" event. In most cases this doesn't matter, Mavlink is forwarded
by mavlink-routerd, OpenHD telemetry/rssi info is forwarded manually,
settings don't require it, and microservices are TCP.

The only thing remaining that will not be forwarded by default is all
of the other flight controller telemetry protocols, like VOT, FrSky,
etc. So in that one case, we manually add the localhost address to
the UDPSplitter in charge of flight controller telemetry, and tell it
to forward that stream to the ground station itself on the usual port.