Read settings file directly in old OSD
The old OSD has always used compiler defines as a form of settings,
with #define statements in the osdconfig.txt file for users to edit.

This requires recompiling the entire thing to change any setting, and
the Open.HD scripts made that worse by only compiling once at boot.

Instead we will just read the settings file and use the values for
if () else () statements.

This has uncovered several compile errors when certain telemetry types
were enabled, which users would not have known how to fix or why the
error occurred. Instead code errors are all caught by the compiler
at build time now, so users will only have to deal with settings file
errors.

Previously, any error in the settings file could have caused the OSD
to fail to compile, leaving users to guess what happened.

Not if there is an error in a users settings file, the OSD will still
run but will use the default value for whichever setting had an error.

In the worst case if everything is corrupted or the entire file is gone
it will still run using all defaults.