Wait for status service in early boot
When the system is first starting, we want the status service to be
running and actually ready to receive events from the rest of the
system so it doesn't miss anything important. If it does, it might
not be visible on the boot screen if something goes wrong.

Instead of sleeping the scripts we will wait for a signal in /tmp,
which tells us that the status service has opened the local UDP port
and is processing messages. This should be done with systemd
soon instead.