Increase video recording space in memory
This increases the space to 1/3 + 30mb, which works out like this:

Pi3a+: (512mb total - 128mb GPU) / 3 = 128, + 30mb = 158mb

Pi3b+: (1024mb total - 128mb GPU) / 3 = 298, + 30mb = 328mb

Pi4b: (2048mb total - 128mb GPU) / 3 = 640, + 30mb = 670mb

At an average bitrate of 5mbit, the pi3a+ could record 5:16, the
pi3b+ could record 10:56, and the 2gb pi4 could record 22:20. A pi4
4gb could record more than anyone is likely to fly.

We should probably disable recording on the pi3a+ so that it
doesn't cause problems, nobody is going to want to record for just 5
minutes anyway.

We're going to add the ability to record to USB sticks and fix the
microsd recording as well but people may still want to record to
ram since it is guaranteed to be reliable and not affect other things,
so long as the recording doesn't cause the rest of the system to run
out of memory.