Prevent throttling on undervolt condition
When the pi detects undervoltage it will reduce the speed of the CPU
and GPU. Generally this is what you want, but in practice there seems
to be some difference between the voltage level where the undervolt
condition is triggered, and the voltage level that would actually
cause a problem.

This is going to require some testing, as we don't want the pi to
actually crash and without throttling that's at least a possibility
if the power supply is very weak.