Enable the FIQ_FSM for pre-pi4 usb system
The Pi zero/ pi3 CPU does not have a normal USB controller on-board,
but one that is really intended for OTG use cases like connecting a
phone to a desktop PC for syncing media files. It does support host
mode, and we're obviously making good use of that, but it does not
work very well in many cases.

To address some of the problems with it, one of the raspberry pi
engineers wrote a bit of kernel code that does a better job handling
certain kinds of USB devices/interactions, and this cmdline parameter
enables it.

It was disabled in Open.HD in the past, likely because it did not work
very well in older kernels, but it is working fine now (and it's the
default on Raspbian anyway).

This change has been tested on pi zero, 4, and pi3b+ with 3x Atheros
cards and various other USB devices connected.