Save to a unique directory on USB by date
This will make it easier to track which video is which, and avoid
conflicts. For the moment the date/time will be wrong, because the
pi has no realtime clock and normally does not have an internet
connection. We will need to set the date/time on the ground station
via mavlink/ltm telemetry from GPS, or by letting QOpenHD on android
send time messages to the ground so it can set itself.

Once we do that the date/time listed on the save directory will be
accurate.