Don't stop hotspot video because of 1 ping fail
When hotspot clients connect, they request an IP address using DHCP
from dnsmasq, which then starts this script as a sort of callback
event.

This script then registers that device with the various telemetry and
video streams and causes the device to start receiving them. It then
periodically sends a ping to the device to see if it's still
connected. However there are cases where a connected device might not
respond to pings quick enough, which would cause the video to stop.

We will now allow devices 10 seconds to reply to each ping, and if 6
pings in a row fail then we consider the device disconnected. In
other words, if the device doesn't respond to pings for a full minute
we stop the video and telemetry streams.

We can afford to continue sending video to devices that are no longer
connected for 60 seconds without causing much trouble, but the
reverse case where a connected device suddenly stops receiving video,
potentially during flight, is not something we can allow to happen.