Add IMX307 imagedir setting
We should be able to autodetect camera models to tell them apart, but
in practice that does not work reliably. Instead we can run the
commands for each one and expect them to fail on the wrong camera. The
commands we are running do not overlap between each model so unless
there are undocumented commands they should not affect the other model.