Switch dependency versions to >=
There is simply no point in any of these version requirements, when
they are set to = things will almost certainly break as soon as a
newer package exists, when they are set to <= the same thing happens.

With >=, apt would grab a newer package if it existed, but it would
have done that anyway without a version requirement specified, which
makes all of this pointless.

We're going to have to track groups of packages externally for qopenhd,
and apt is not capable of doing it anyway.