Allow Mavlink compid 200 in old OSD
Ardupilot does the right thing with sysid 1 and the compid constant
for an autopilot. iNav uses sysid 1 + compid 250 for some reason, and
betaflight uses sysid 0 + compid 200.

Because the OSD is read only, we don't need to worry about which
compid to send messages to, we can still filter this way.