Reject duplicate hotspot addresses in udpsplitter
Now that the removal timeout is an entire minute, it's possible for
someone to join 2+ times before the timeout would have occurred, which
makes it possible for the same IP to end up in the list multiple times,
and that will cause horribly broken telemetry/video to be sent to
hotspot devices.

So we check first to see if it's already in the list, and skip it
if so.