Override vm.min_free_kbytes in raspberrypi-sys-mod
The upstream package contains a file called 98-rpi.conf, which
has a vm.min_free_kbytes line that breaks on pi3a+ for some reason,
we should not touch that file but we can override it with a file that
should be guaranteed to be evaluated last.

The reason vm.min_free_kbytes is an issue is not clear yet, but the
pi3a+ is acting like it has very little memory even when it clearly
has plenty. Stretch does not show this problem, so it may be a kernel
bug of some kind.