Remove speed trend and yaw trend from osd
It is supposed to show trends. The altitude one works correctly. 
The speed and yaw have bugs because the mavlink messages they read aer in a global motion context, instead of local. 

The fix is to convert the global x y motion vectors to a local context x y motion vector.