Wait for gnuplot before unmounting USB drive
When the video save system is running it needs to complete all of the
individual tasks before unmounting and telling the user they can
remove the USB drive.

At the moment that was not happening, gnuplot was being allowed to run
in the background, which created an opportunity for the drive to be
unmounted before the plots could be generated and saved.