Wait for eth0 to change state to up, don't race it
This is a race condition and it frequently loses at the moment,
resulting in the ethernet interface not being up before we try to
configure it.

Instead of sleeping for an arbitrary time, we will ask the interface
whether it is up or not and wait until it is before proceeding.

This should make it reliable.