Remove the old joystick quirk setting + code
This was intended to fix the Geekpi joystick board glitches, where
the proportional channels would spontaneously and repeatedly reset
to the lowest possible RC value, 1000. This was dangerous, so Thomas
found a workaround that prevented the problem from affecting flight.

However, the real cause of those problems was a USB timing issue on
the older Pi models, which have now been resolved directly by using
fiq_fsm, so this code is no longer necessary.