Add system manifest service
This runs at early startup, long before anything else is ready, and
does platform, camera, and wifi card autodetection. It then generates
manifest files for each of those classes of device and stores them
so other system components can use them.

This is a one-shot service for now, but it could easily rescan
periodically and signal the rest of the system to re-read the
manifests, for example to support camera hotplug, or wifi card
reconnections.

We will likely need to ensure the USB devices are all enumerated by
the kernel before this runs, but otherwise there aren't any
interactions with the rest of the system.