Add security service
This is primarily responsible for key handling, for the moment that
means deriving a set of keypairs from a user-provided passphrase,
however we won't be doing that much longer.

This service will also be wired up to the microservice channel and
used to allow key backup and selective sharing, changing the hotspot
password at runtime, etc.