Remove openhdvid, replacing with camera service
This was a raspivid clone built around Picamera, however it never
worked quite right due to buffering issues, it was difficult to
extend in the camera side because all of the MMAL code was hidden
away in the python module, and it's completely non-portable it only
works on the Pi. It's also based around MMAL, which isn't going to
be used on the Pi for much longer due to the pi engineers switching
toward standard linux interfaces.