Add dependency on libnl
This is used for directly interacting with netlink APIs for cases that
should not be implemented as a mess of shell commands and regexes.

Determining whether a particular wifi card actually supports 2g and 5g
bands is one initial use (with the caveat that some 5g cards don't
have 2g RF front ends, but we can special case those).