Include air/ground designation in profile manifest
This replaces the various duplicated parts of the old scripts that
attempted to use platform-specific checks to determine whether the
system was running as an air unit or a ground unit by looking for
a pi camera.

This makes a single determination once at boot, then at runtime any
services only need to check the profile. This also makes the check
generic across different kinds of cameras, and cross platform.

This removes any need for people to create a /boot/air.txt file.