Add interface service
This service is responsible for all WiFi, SDR, Ethernet, LTE and other
transmission or hotspot interfaces.

It reads the profile manifest and the wifi/ethernet/lte manifests to
determine which ones need to be configured upfront in order to support
basic function, and which ones require further input from the user
via the settings system.

The WiFi handling code is present in this commit but still needs a
few small changes, primarily to support the microservice interface
for turning the hotspot on and off, setting the txpower live, and
for reconfiguring the cards live to designate one of them as a hotspot
device if it isn't obvious from the hardware configuration.

For now, a card that is known not to support or work well for
injection is chosen for the hotspot by default.

This code also needs to use the systemd APIs to signal when the
hotspot interface is ready to use, so that it can start hostapd and
dnsmasq.