Add ethernet and tethering types
In most cases we don't care what kind of ethernet interface it is,
but iPhone and android devices must be detected and marked
accordingly.

If another ethernet device is found, by default it should be used for
ethernet hotspot purposes (the interface service handles this),
unless the system is running as an air unit.