Add common microservice base class
This is primarily derived from the openhd-microservice code which
is going to use this common base class now, as do several of the
newer components.

This is a simplistic implementation of command handling, it's only
intended to be used for UI commands or single commands sent
automatically by the system itself, rather than sequences of steps for
complex actions.

This code is not intended to handle anything related to the flight
controller, it does not need much more than what is already here.

Still, we might replace the manual code with one of the mavlink SDKs
if one of them is a good fit.