Remove old raspi2raspi binary
No source in the tree so it can't stay here anyway, but it's not
something we can build features around anymore because it only works
on raspberry pi.

QOpenHD can do display mirroring cross-platform, we just haven't
enabled it as a feature because we still need to turn on zero-copy
rendering of video inside the app.