Move and preserve the old tracker code
Also adds a makefile for it to allow easier building and testing,
should anyone want to use it in the future.

For now we don't ship a tracker in the image itself. This one was
tied to the raspberry pi anyway, and most people use external trackers
like ghettostation now. External trackers are much easier to deal with,
and it keeps openhd itself from being even more complicated than it
already is.