Remove the old main.sh script
This is where all the ttys were starting from, however we don't do
that anymore we just have systemd units that start isolated services
with well defined interfaces between them, we have no reason to run
stuff this way anymore.

All of the scripts that remain from this system that are still around
are going to be called from individual services if needed.