Add early_init script for temporary commands
The sole reason this exists at the moment is because we haven't yet
enabled zero-copy rendering in QOpenHD, and we haven't dropped the
old OSD. The hello_video binary is not portable to any other platform
so it makes no sense to build things around it anymore. We could use
gstreamer temporarily if we need to.

We may also need to run a few other things here. It should be started
from systemd right after the early target, and the next component
in the startup chain should wait for it.