Allow openhd-system to find Jetson CSI sensors
This is the raw sensor side of the camera pipeline, so it will find
v4l2 devices that claim to only provide raw bayer data and not
h264/h265, which is true as they are just sensors and this is how
the Jetson platform works.

The camera microservice will tell the argus camera system which
sensor to use, and then it takes care of setting up the ISP and
encoder.

The camera manifest won't show these sensors to be h264 or h265
capable but that will not be visible to users, they will instead see
high level settings from the camera service that will properly show
h264 and h265 capability for the full sensor+ISP+encoder pipeline.