Remove the width/height/fps caps on SEEK pipelines
The v4l2loopback setup the seek uses does not allow negotiation of
the resolution or framerate, it is very simple and requires the
v4l2 client to accept whatever comes through the device. It does
however have correct info about the resolution and framerate in use,
because seek_viewer sets it properly with v4l2 APIs. As a result we
can just leave this out.