Auto calibrate Seek Compact at power on
When the drone is powering up, place something flat like a piece of
paper (not reflective) in front of the Seek camera until the thermal
video appears on the OSD.

This will calibrate the sensor and eliminate a lot of noise from the
image.

This is not a permanent calibration, it needs to be done each time
the camera is turned on.