Add camera settings template
This is not the actual settings file but a template used to generate
a camera section in the camera settings file. There can be more than
one section, and if settings are added or removed this template
system ensures that users are not left behind trying to figure out
what changed.

This works exactly the same way qopenhd.conf works already.