Add inja template library
Used for settings fragment templates. Each piece of hardware that is
discovered and written to a settings file, or manually added to a
settings file, is generated from one of the template files which are
to be used with the inja library.

This allows them to be updated automatically so that users don't
end up with outdated settings files. QOpenHD works the same way.