Automatically configure wifi hotspot hostapd/dhcp
This does not do the old opposite band trick, because it seems in
practice people prefer usb or ethernet instead of 2.4Ghz wifi hotspot,
it may even be the case that 5Ghz wifi hotspot would work better even
when the transmission cards are also 5Ghz.

We can add the opposite band trick back in very easily if we need to,
because the Card types we process elsewhere in this class know whether
any of the wifibroadcast cards are using a particular band.