Add support for item types in settings parser
They don't need to be tracked by name because the context the file
is used in tells the system which item type will be found, but they
still need to be separated cleanly so users know what they are.

For example the camera.conf file can have several [Camera] items,
one for each camera that is currently or has been connected to a
vehicle at some point. On first boot if a camera is found that does
not have an item in the settings file, it will be added automatically
with the default settings for that camera.

Users can also choose to add an item manually even if the hardware
it describes has never been connected before. This means settings for
different wifi cards or cameras can be preserved without duplicating
entire profiles or making people manually edit things to change the
system behavior.