Fix Ralink rssi display
This is a crude hack, but it will work until we replace everything
in a few weeks anyway.

The issue is that wifi cards generally have multiple antenna ports,
but which one actually has an antenna connected is different on
different cards, sometimes even cards with the same chip will have
different antenna arrangements.

On most cards it works fine to only look at the first antenna index,
but even that is not really correct and totally ignores the fact that
there may not be only one antenna.

Ralink cards happen to be using antenna index 1 instead of 0, at least
those that I have tested.

Ref: #220