Include camera indexes in manifest
This is the discovery order, which becomes the camera order by
default, which users can change by setting the indexes in camera.conf.

For compatibility with existing apps like qgc and fpv_vr, users with
multiple cameras connected can change the index of their main camera
to index 0 in camera.conf to ensure it is selected as the main
stream.