Load and safe wifi card settings on air/ground
This is the local part of the wifi card setting system, which allows
individual wifi cards to be configured for a specific purpose, such
as wifibroadcast, hotspot, relay, or wifi client, and to
have their own frequency and txpower settings.

There is no initial setup required, even the settings file isn't
required. If you connect a wifi card that supports wifibroadcast, it
will be automatically used that way. If it only supports hotspot use,
it will be configured as a hotspot. You can change that in the settings
file after the system boots for the first time.

On the ground side the wifi card settings are in this file:

/conf/openhd/ground/wifi.conf (this is the /boot partition on platforms
that have one, otherwise it's a dedicated config file partition).

There is one entry for each card. If you decide to connect a 2nd wifi
card to an existing system you have been using, it will automatically
show up in the file the next time the system boots, without affecting
the settings for the original wifi card.

Vehicle settings are stored in their own directory, separately:

/conf/openhd/vehicle1/wifi.conf

Users can rename the vehicle folders if they wish, the system will
still know which one is which.

The next piece of this system that is not merged yet is the sync
handling code, which will allow users to edit settings for each
vehicle on the ground SD card, and automatically sync the settings
to the vehicle the next time it connects, without requiring a
specific RC/GPIO trigger like SmartSync did.