Load and save ethernet settings in interface svc
This works the same way the wifi settings do, ethernet interfaces
found on the system (or interfaces that pretend to be ethernet) are
autoconfigured for static, hotspot, or client mode and if there is
a settings entry for the interface those settings are respected.

Once configured the settings that were actually used get written back
out to the settings file.