Add telemetry config template
These may be entirely unnecessary soon, even the serial port doesn't
really need to be configurable as we know what it will be on each
platform. People who want to connect USB serial adapters can do that
and we can automate it rather than making them mess with settings.

Even the telemetry protocol can be autodetected very easily, but for
the moment we will make it explicit until the system is well tested.