Add telemetry service
This replaces the older mavlink-router and OpenHDRouter code, which
were only capable of routing Mavlink and did not integrate with the
settings system or the platform manifest system.

This code is designed to allow for other protocols such as MSPv2, LTM,
FrSky, etc. Those protocols do not actually have an addressing system,
however we only need to forward them out of every endpoint except the
one the message arrived on (so don't send messages right back to the
flight controller).

This will enable things like 2-way configuration for Betaflight/iNav,
as well as MSPv2 RC protocol support.

We may end up automating the telemetry type making that setting
unnecessary, and we may also automatically add USB serial converters
to the endpoint list making it unnecessary for the serial ports to be
configurable either.

There are still 2 things left to do in this code:

1) Add parsers for the other telemetry protocols (trivial)

We just need it to find message boundaries and forward complete
messages as-is. We already have code that can handle this in QOpenHD,
it should be reused here.

2) React to DHCP client join/leave events (30 minutes of work)

The mavlink-router code we used before had changes added to do this,
it just requires listening on a specific UDP port and adding UDP
endpoints dynamically so that messages get sent to hotspot clients
without making people manually add endpoints to the settings file.