Store UDP endpoint addresses for dynamic management
When the service is sent a command to add a new endpoint on a hotspot
or other connected device we need to be able to find an existing
endpoint so we don't add it twice, and so we can remove them easily.

So we store the address in string format for easy comparison.