Add video service
This service is the video handling code for the entire system, it
discovers and configures cameras, handles video routing to and from
wifibroadcast adapters, handles routing video to hotspot devices and
the local system, and handles recording triggered by Mavlink.

The recording code is not being merged with this yet, and Mavlink is
not enabled in the code but the code itself supports it already.